fix(pwa): restrict trusted auth preview origins
Summary:
- Replaced provider-wide Workers and Pages auth/CORS trust with the
  account-scoped *.horusdev.workers.dev pattern.
- Moved auth origin matching into a testable helper module.
- Added API-origin tests that reject unrelated workers.dev and pages.dev
  origins.
- Included API tests in the PWA Vitest config.

Rationale:
- Credentialed CORS with SameSite=None cookies must not trust arbitrary
  Cloudflare provider preview domains.

Tests:
- vp check --fix packages/pwa/api/auth.ts packages/pwa/api/auth-origins.ts packages/pwa/api/auth-origins.test.ts packages/pwa/wrangler.jsonc packages/pwa/vite.config.ts
- vp run --filter @trizum/pwa test
- vp run --filter @trizum/pwa check
- vp run --filter @trizum/pwa build
