+++ packages/pwa/api/types.d.ts
@@ -1,5 +1,5 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types --path=./api/types.d.ts` (hash: c8ee9cea650c2f16f92bde66a75d08bb)
+// Generated by Wrangler by running `wrangler types --path=./api/types.d.ts` (hash: 6668965ba6e07ffd88008e78856a5697)
 // Runtime types generated with workerd@1.20260312.1 2025-11-18 nodejs_als,nodejs_compat
 declare namespace Cloudflare {
 	interface GlobalProps {
@@ -12,8 +12,8 @@ declare namespace Cloudflare {
 		APPLE_APP_BUNDLE_IDENTIFIER: "app.trizum.capacitor";
 		APPLE_SERVICE_ID: "app.trizum.capacitor.si";
 		AUTH_EMAIL_FROM: "trizum <noreply@trizum.app>";
-		BETTER_AUTH_ALLOWED_HOSTS: "trizum.app,*.workers.dev,*.pages.dev,localhost:5173,localhost:8787,127.0.0.1:5173,127.0.0.1:8787";
-		BETTER_AUTH_TRUSTED_ORIGINS: "https://trizum.app,https://*.workers.dev,https://*.pages.dev,capacitor://localhost,ionic://localhost,http://localhost,http://localhost:5173,http://localhost:8787";
+		BETTER_AUTH_ALLOWED_HOSTS: "trizum.app,*.horusdev.workers.dev,localhost:5173,localhost:8787,127.0.0.1:5173,127.0.0.1:8787";
+		BETTER_AUTH_TRUSTED_ORIGINS: "https://trizum.app,https://*.horusdev.workers.dev,capacitor://localhost,ionic://localhost,http://localhost,http://localhost:5173,http://localhost:8787";
 		BETTER_AUTH_URL: "https://trizum.app";
 		GOOGLE_ANDROID_APK_CLIENT_ID: "554871818976-9el2tt5b2fj1quodn14be8jncpefij7v.apps.googleusercontent.com";
 		GOOGLE_ANDROID_GOOGLE_PLAY_CLIENT_ID: "554871818976-bps9ha7q5qhop0i8nt8jfa2l2mu48a6a.apps.googleusercontent.com";
