+++ packages/pwa/wrangler.jsonc
@@ -36,8 +36,8 @@
     "APPLE_APP_BUNDLE_IDENTIFIER": "app.trizum.capacitor",
     "APPLE_SERVICE_ID": "app.trizum.capacitor.si",
     "AUTH_EMAIL_FROM": "trizum <noreply@trizum.app>",
-    "BETTER_AUTH_ALLOWED_HOSTS": "trizum.app,*.workers.dev,*.pages.dev,localhost:5173,localhost:8787,127.0.0.1:5173,127.0.0.1:8787",
-    "BETTER_AUTH_TRUSTED_ORIGINS": "https://trizum.app,https://*.workers.dev,https://*.pages.dev,capacitor://localhost,ionic://localhost,http://localhost,http://localhost:5173,http://localhost:8787",
+    "BETTER_AUTH_ALLOWED_HOSTS": "trizum.app,*.horusdev.workers.dev,localhost:5173,localhost:8787,127.0.0.1:5173,127.0.0.1:8787",
+    "BETTER_AUTH_TRUSTED_ORIGINS": "https://trizum.app,https://*.horusdev.workers.dev,capacitor://localhost,ionic://localhost,http://localhost,http://localhost:5173,http://localhost:8787",
     "BETTER_AUTH_URL": "https://trizum.app",
     "GOOGLE_ANDROID_APK_CLIENT_ID": "554871818976-9el2tt5b2fj1quodn14be8jncpefij7v.apps.googleusercontent.com",
     "GOOGLE_ANDROID_GOOGLE_PLAY_CLIENT_ID": "554871818976-bps9ha7q5qhop0i8nt8jfa2l2mu48a6a.apps.googleusercontent.com",
